--- pyveda/frameworks/pretrainedmodels/keras/classification/__init__.py
@@ -1,14 +1,0 @@
-try:
-    import keras
-except ImportError:
-    raise ImportError('Keras not installed. Please install keras to import models')
-
-
-from pyveda.frameworks.pretrainedmodels.keras.classification.resnet50 import ResNet50
-from pyveda.frameworks.pretrainedmodels.keras.classification.vgg16 import VGG16
-from pyveda.frameworks.pretrainedmodels.keras.classification.vgg19 import VGG19
-from pyveda.frameworks.pretrainedmodels.keras.classification.inception_v3 import InceptionV3
-from pyveda.frameworks.pretrainedmodels.keras.classification.densenet121 import DenseNet121
-from pyveda.frameworks.pretrainedmodels.keras.classification.xception import Xception
-from pyveda.frameworks.pretrainedmodels.keras.classification.nasnet import NASNetLarge
-from pyveda.frameworks.pretrainedmodels.keras.classification.inception_resnet_v2 import InceptionResNetV2
