--- pyveda/frameworks/pretrainedmodels/keras/classification/densenet121.py
@@ -1,3 +1,0 @@
-def DenseNet121(**kwargs):
-    from keras.applications.densenet import DenseNet121
-    return DenseNet121(**kwargs)
