--- pyveda/frameworks/pretrainedmodels/keras/classification/inception_resnet_v2.py
@@ -1,3 +1,0 @@
-def InceptionResNetV2(**kwargs):
-    from keras.applications.inception_resnet_v2 import InceptionResNetV2
-    return InceptionResNetV2(**kwargs)
