--- pyveda/frameworks/pretrainedmodels/keras/classification/resnet50.py
@@ -1,4 +1,0 @@
-def ResNet50(**kwargs):
-    #assert has_keras, 'To use this model install keras'
-    from keras.applications.resnet50 import ResNet50
-    return ResNet50(**kwargs)
