--- pyveda/frameworks/pretrainedmodels/keras/classification/xception.py
@@ -1,4 +1,0 @@
-def Xception(**kwargs):
-    from keras.applications.xcepetion import Xception
-    return Xception(**kwargs)
-
