--- pyveda/frameworks/pretrainedmodels/pytorch/classification/alexnet.py
@@ -1,3 +1,0 @@
-def AlexNet(**kwargs):
-    import torchvision.models as models
-    return models.alexnet(pretrained = True, **kwargs)
