--- pyveda/frameworks/pretrainedmodels/pytorch/classification/densenet161.py
@@ -1,3 +1,0 @@
-def DenseNet161(**kwargs):
-    import torchvision.models as models
-    return models.densenet161(pretrained = True, **kwargs)
