--- pyveda/frameworks/pretrainedmodels/pytorch/classification/inception_v3.py
@@ -1,3 +1,0 @@
-def InceptionV3(**kwargs):
-    import torchvision.models as models
-    return models.inception_v3(pretrained = True, **kwargs)
