--- pyveda/frameworks/pretrainedmodels/pytorch/classification/resnet18.py
@@ -1,3 +1,0 @@
-def ResNet18(**kwargs):
-    import torchvision.models as models
-    return models.resnet18(pretrained = True, **kwargs)
