--- pyveda/frameworks/pretrainedmodels/pytorch/classification/resnet50.py
@@ -1,3 +1,0 @@
-def ResNet50(**kwargs):
-    import torchvision.models as models
-    return models.resnet50(pretrained = True, **kwargs)
