--- pyveda/frameworks/pretrainedmodels/pytorch/classification/squeezenet.py
@@ -1,3 +1,0 @@
-def SqueezeNet(**kwargs):
-    import torchvision.models as models
-    return torchvision.models.squeezenet(pretrained = True, **kwargs)
