--- pyveda/frameworks/pretrainedmodels/pytorch/classification/vgg16.py
@@ -1,3 +1,0 @@
-def VGG16(**kwargs):
-    import torchvision.models as models
-    return models.vgg16(pretrained = True, **kwargs)
